requirements.txt
appium-python-client>=2.1.0
pytest>=7.0.0
allure-pytest>=2.9.0
selenium>=4.0.0